+++ ports/atmel-samd/boards/circuitplayground_express/mpconfigboard.h
@@ -38,7 +38,7 @@
 
 #define SPEAKER_ENABLE_PIN    (&pin_PA30)
 
-#include "spi_flash.h"
+#include "external_flash.h"
 
 // If you change this, then make sure to update the linker scripts as well to
 // make sure you don't overwrite code.
