+++ ports/atmel-samd/boards/feather_m0_supersized/mpconfigboard.h
@@ -33,7 +33,7 @@
 #define MICROPY_PORT_B        ( 0 )
 #define MICROPY_PORT_C        ( 0 )
 
-#include "spi_flash.h"
+#include "external_flash.h"
 
 // If you change this, then make sure to update the linker scripts as well to
 // make sure you don't overwrite code.
