+++ ports/atmel-samd/boards/metro_m4_express/mpconfigboard.h
@@ -40,7 +40,7 @@
 
 #define AUTORESET_DELAY_MS 500
 
-#include "spi_flash.h"
+#include "external_flash.h"
 
 // If you change this, then make sure to update the linker scripts as well to
 // make sure you don't overwrite code
