+++ ports/atmel-samd/boards/trinket_m0_haxpress/mpconfigboard.h
@@ -40,7 +40,7 @@
 
 #define CALIBRATE_CRYSTALLESS 1
 
-#include "spi_flash.h"
+#include "external_flash.h"
 
 // If you change this, then make sure to update the linker scripts as well to
 // make sure you don't overwrite code.
@@ -49,4 +49,3 @@
 #define BOARD_FLASH_SIZE (0x00040000 - 0x2000 - CIRCUITPY_INTERNAL_NVM_SIZE)
 
 #include "flash_W25Q32BV.h"
-
