+++ ports/atmel-samd/boards/ugame10/mpconfigboard.h
@@ -33,7 +33,7 @@
 
 #define CALIBRATE_CRYSTALLESS 1
 
-#include "spi_flash.h"
+#include "external_flash.h"
 
 // If you change this, then make sure to update the linker scripts as well to
 // make sure you don't overwrite code.
