+++ ports/atmel-samd/external_flash.h
@@ -23,8 +23,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-#ifndef MICROPY_INCLUDED_ATMEL_SAMD_SPI_FLASH_H
-#define MICROPY_INCLUDED_ATMEL_SAMD_SPI_FLASH_H
+#ifndef MICROPY_INCLUDED_ATMEL_SAMD_EXTERNAL_FLASH_H
+#define MICROPY_INCLUDED_ATMEL_SAMD_EXTERNAL_FLASH_H
 
 #include <stdbool.h>
 
@@ -40,21 +40,21 @@
 #define SPI_FLASH_SYSTICK_MASK    (0x1ff) // 512ms
 #define SPI_FLASH_IDLE_TICK(tick) (((tick) & SPI_FLASH_SYSTICK_MASK) == 2)
 
-void spi_flash_init(void);
-uint32_t spi_flash_get_block_size(void);
-uint32_t spi_flash_get_block_count(void);
-void spi_flash_irq_handler(void);
-void spi_flash_flush(void);
-bool spi_flash_read_block(uint8_t *dest, uint32_t block);
-bool spi_flash_write_block(const uint8_t *src, uint32_t block);
+void external_flash_init(void);
+uint32_t external_flash_get_block_size(void);
+uint32_t external_flash_get_block_count(void);
+void external_flash_irq_handler(void);
+void external_flash_flush(void);
+bool external_flash_read_block(uint8_t *dest, uint32_t block);
+bool external_flash_write_block(const uint8_t *src, uint32_t block);
 
 // these return 0 on success, non-zero on error
-mp_uint_t spi_flash_read_blocks(uint8_t *dest, uint32_t block_num, uint32_t num_blocks);
-mp_uint_t spi_flash_write_blocks(const uint8_t *src, uint32_t block_num, uint32_t num_blocks);
+mp_uint_t external_flash_read_blocks(uint8_t *dest, uint32_t block_num, uint32_t num_blocks);
+mp_uint_t external_flash_write_blocks(const uint8_t *src, uint32_t block_num, uint32_t num_blocks);
 
-extern const struct _mp_obj_type_t spi_flash_type;
+extern const struct _mp_obj_type_t external_flash_type;
 
 struct _fs_user_mount_t;
 void flash_init_vfs(struct _fs_user_mount_t *vfs);
 
-#endif  // MICROPY_INCLUDED_ATMEL_SAMD_SPI_FLASH_H
+#endif  // MICROPY_INCLUDED_ATMEL_SAMD_EXTERNAL_FLASH_H
